aag/data_pipeline/data_transformer/text_loader.py
"""
text_loader.py
----------------
文本数据加载引擎（TextDataLoader），未实现

负责：
1. 从 YAML 注册表读取文本数据集定义；
2. 加载文本文件（txt / json / markdown / docx 等）；
3. 转换为文档对象列表（list[dict]）；
"""

from typing import Any, Dict, List, Optional
import yaml
from aag.config.data_upload_config import *


class TextDataLoader:
    """文本数据加载引擎，用于加载非结构化文本类数据集。"""

    def __init__(self, registry_path: str):
        """
        初始化文本数据加载器。

        Args:
            registry_path (str): 文本数据注册文件路径（text_registry.yaml）
        """
        self.registry_path = registry_path
        self.dataset_schemas: Dict[str, DatasetConfig] = {}   
        pass

    def _load_registry(self) -> Dict[str, Any]:
        """
        加载文本数据注册表（YAML 文件）。

        Returns:
            dict: 已解析的注册表内容。
        """
        pass

    # ==========================================================
    # 主功能接口
    # ==========================================================

    def list_datasets(self) -> List[str]:
        """
        列出当前注册表中所有文本数据集名称。
        """
        # TODO: 遍历 self.registry_data["datasets"] 并返回所有 name 列表
        return []

    def get_dataset_info(self, name: str) -> Dict[str, Any]:
        """
        获取指定文本数据集的详细配置信息。

        Args:
            name (str): 数据集名称。
        Returns:
            dict: 数据集定义信息。
        """
        # TODO: 查找 self.registry_data["datasets"] 中 name == 指定名称 的项
        return {}

    def load_text_dataset(self, name: str) -> List[Dict[str, Any]]:
        """
        加载指定名称的文本数据集。

        Args:
            name (str): 数据集名称。
        Returns:
            list[dict]: 每条记录为 {"id": str, "text": str, "meta": dict}
        """
        # TODO:
        # 1. 查找数据集定义
        # 2. 根据 format 类型 (txt/json/md/docx等) 选择读取方式
        # 3. 解析文本文件，组织成标准格式 [{"id": ..., "text": ..., "meta": {...}}]
        # 4. 返回结果
        return []

    # ==========================================================
    # 可扩展方法
    # ==========================================================

    def preview(self, name: str, n: int = 3) -> Optional[List[str]]:
        """
        预览前 n 条文本内容。
        """
        # TODO: 调用 load_text_dataset(name)，返回前 n 条的 text 字段
        return None

    def to_documents(self, name: str) -> List[Dict[str, Any]]:
        """
        转换为统一文档对象格式，用于后续向量化。
        """
        # TODO:
        # 1. 加载文本
        # 2. 结构化成 {"id": ..., "text": ..., "source": ...}
        # 3. 返回标准化的文档结构
        return []

    def register_new_dataset(self, dataset_info: Dict[str, Any]) -> None:
        """
        将新的文本数据集注册到 YAML 文件中。

        Args:
            dataset_info (dict): 数据集配置项。
        """
        # TODO:
        # 1. 加载原 YAML
        # 2. 追加 dataset_info
        # 3. 保存回注册表文件
        pass

    def validate_schema(self, dataset_info: Dict[str, Any]) -> bool:
        """
        校验文本数据集的 schema 是否符合模板。
        """
        # TODO: 加载 schema/text_schema_template.yaml 进行字段校验
        return True
